jlist changed to jtable and reloading of the lists
The list of the files to tag needs to be reloaded every time the user changes a tag. In the future this will work, for now it has some issues. The jlist that contained the files to tag has been changed to a jtable which in the future will contain more information about the files.